Use CSV lib to parse and join csv lines when passing values to given block
When there are commas inside fields and block is given there is a problem with `row`.
Line is split by delimiter, and when delimiter is a comma, `row` has more values than there are columns.

Using CSV lib to parse the line avoids splitting just by delimiter. Line is split with regards to CSV format.
When `yield` finishes, line is again generated by CSV lib `generate_line` to ensure the line is a valid CSV line.